// frontend/src/api-client/models/AuthUserDto.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type AuthUserDto = {
    id?: string;
    email?: string;
    fullNameVi?: string;
    roles?: Array<string>;
    tenantId?: string;
    branchId?: string;
};