// frontend/src/api-client/models/LoginResponse.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
import type { AuthUserDto } from './AuthUserDto';
export type LoginResponse = {
    token?: string;
    refreshToken?: string;
    expiresAt?: string;
    user?: AuthUserDto;
};